refactor: Validation 로직을 model에서 view로 이동
