+++ app/src/main/java/nextstep/signup/ui/signup/SignUpScreen.kt
@@ -16,10 +16,10 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import nextstep.signup.R
-import nextstep.signup.model.validation.CompositeValidation
-import nextstep.signup.model.validation.EqualValidation
-import nextstep.signup.model.validation.LengthValidation
-import nextstep.signup.model.validation.RegexValidation
+import nextstep.signup.ui.validation.CompositeValidation
+import nextstep.signup.ui.validation.EqualValidation
+import nextstep.signup.ui.validation.LengthValidation
+import nextstep.signup.ui.validation.RegexValidation
 import nextstep.signup.ui.theme.SignupTheme
 
 @Composable
