+++ app/src/main/java/nextstep/signup/ui/validation/CompositeValidation.kt
@@ -1,4 +1,4 @@
-package nextstep.signup.model.validation
+package nextstep.signup.ui.validation
 
 class CompositeValidation(
     private vararg val validations: Validation
