+++ app/src/main/java/nextstep/signup/ui/validation/LengthValidation.kt
@@ -1,11 +1,11 @@
-package nextstep.signup.model.validation
+package nextstep.signup.ui.validation
 
-class RegexValidation(
-    private val regex: Regex,
+class LengthValidation(
+    private val range: IntRange,
     private val errorMessage: String,
 ) : Validation {
     override fun validate(text: String): Boolean =
-        regex.matches(text)
+        text.length in range
 
     override fun errorMessage(text: String): String = errorMessage
 }
