+++ app/src/main/java/nextstep/signup/ui/validation/Validation.kt
@@ -1,4 +1,4 @@
-package nextstep.signup.model.validation
+package nextstep.signup.ui.validation
 
 interface Validation {
     fun validate(text: String): Boolean
