+++ app/src/test/java/nextstep/signup/EmailValidationTest.kt
@@ -2,10 +2,9 @@ package nextstep.signup
 
 import nextstep.signup.ui.signup.EmailTextField
 import androidx.compose.runtime.mutableStateOf
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
-import nextstep.signup.model.validation.RegexValidation
+import nextstep.signup.ui.validation.RegexValidation
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
